# api/openapi.yaml
openapi: 3.1.0
info:
  title: OnboardMePlease API
  version: 0.4.0
  description: >-
    Capture, resumable synthesis, searchable OKF knowledge and source-grounded questions are implemented.
    Deletion and separate snapshot creation remain planned; those routes are marked below.
servers:
  - url: http://127.0.0.1:8765
    description: Loopback default for the local single-user service.
paths:
...
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview:
    get:
      operationId: getOverview
      summary: Read the current technical onboarding overview
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      responses:
        '200':
          description: Overview and limitations
          content:
            application/json:
              schema: { $ref: '../contracts/analysis.schema.json#/$defs/Overview' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/inventory:
    get:
      operationId: getOverviewInventory
      summary: Page through every accounted-for snapshot artifact
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: offset
          in: query
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '200': { description: Artifact status, category, and reason codes }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/notes:
    post:
      operationId: addOverviewNote
      summary: Retain a local correction attached to a current overview claim
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [claim_id, note]
              properties:
                claim_id: { type: string }
                note: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201': { description: Correction retained without automatic human verification }
        '422': { $ref: '#/components/responses/Problem' }
  /v1/repositories/{repositoryId}/snapshots/{snapshotId}/overview/notes/{noteId}:
    delete:
      operationId: deleteOverviewNote
      summary: Remove one local correction from the selected snapshot
      parameters:
        - $ref: '#/components/parameters/RepositoryId'
        - $ref: '#/components/parameters/SnapshotId'
        - name: noteId
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204': { description: Correction removed }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
    Problem:
      description: Safe error without raw input, code, path, or credential value
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }